polymarket-mcp/requirements.txt
mcp>=1.0.0
py-clob-client
python-dotenv
